chore: move startup runtime to shared infrastructure
