+++ src/app/shared/infrastructure/runtime/run-startup-runtime.ts
@@ -1,4 +1,4 @@
-import { startupRuntime } from '@/app/setup/application/startup-runtime';
+import { startupRuntime } from '@/app/shared/infrastructure/runtime/startup-runtime';
 import { setupSuperAdmin } from '@/composition/setup/super-admin-setup';
 import { mongoDBClient } from '@/composition/shared/persistence/mongodb';
 import { setupConfig } from '@/config/setup';
